feat(hydra-common): move DatabaseManager to hydra-common to minimize the dependencies
affects: @dzlzv/hydra-cli, @dzlzv/hydra-common, @dzlzv/hydra-db-utils, @dzlzv/hydra-indexer,
@dzlzv/hydra-processor, sample, sample-mappings

BREAKING CHANGE:
All DatabaseManager imports should use @dzlzv/hydra-common
