+++ packages/hydra-common/src/db/DeepPartial.ts
@@ -1,0 +1,8 @@
+// extracted from typeorm
+export type DeepPartial<T> = {
+  [P in keyof T]?: T[P] extends Array<infer U>
+    ? Array<DeepPartial<U>>
+    : T[P] extends ReadonlyArray<infer U>
+    ? ReadonlyArray<DeepPartial<U>>
+    : DeepPartial<T[P]> | T[P]
+}
