+++ packages/hydra-common/src/index.ts
@@ -1,2 +1,3 @@
 export * from './interfaces'
 export * from './utils'
+export * from './db'
