+++ packages/hydra-db-utils/src/db/index.ts
@@ -1,3 +1,2 @@
-export * from './DatabaseManager'
 export * from './helper'
-export { SnakeNamingStrategy } from './SnakeNamingStrategy'
+export * from './SnakeNamingStrategy'
