+++ packages/hydra-db-utils/src/index.ts
@@ -1,2 +1,2 @@
-export * from './db'
 export * from './redis'
+export * from './db'
