+++ packages/hydra-processor/src/start/QueryEventProcessingPack.ts
@@ -1,7 +1,5 @@
 // @ts-check
-
-import { DatabaseManager } from '@dzlzv/hydra-db-utils'
-import { SubstrateEvent } from '@dzlzv/hydra-common'
+import { DatabaseManager, SubstrateEvent } from '@dzlzv/hydra-common'
 
 export type QueryEventProcessorResult = void | Promise<void>
 export type EventHandlerFunc = (
